requirements.txt
pandas==2.2.3
python_fasthtml==0.12.18
